--- lib/hausaufgabenheft_logik/lib/src/completed_homeworks/completed_homeworks_view_bloc/completed_homeworks_view_bloc.dart
@@ -1,15 +1,0 @@
-// Copyright (c) 2022 Sharezone UG (haftungsbeschränkt)
-// Licensed under the EUPL-1.2-or-later.
-//
-// You may obtain a copy of the Licence at:
-// https://joinup.ec.europa.eu/software/page/eupl
-//
-// SPDX-License-Identifier: EUPL-1.2
-
-import 'events.dart';
-import 'states.dart';
-
-abstract class CompletedHomeworksViewBloc
-    implements
-        Stream<CompletedHomeworksViewBlocState>,
-        Sink<CompletedHomeworksViewBlocEvent> {}
