+++ lib/hausaufgabenheft_logik/lib/src/completed_homeworks/completed_homeworks_view_bloc/completed_homeworks_view_bloc_impl.dart
@@ -10,22 +10,27 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:bloc_base/bloc_base.dart';
-import 'package:rxdart/rxdart.dart';
-import 'package:hausaufgabenheft_logik/src/completed_homeworks/lazy_loading_completed_homeworks_bloc/events.dart'
-    as lazy_loading;
-import 'package:hausaufgabenheft_logik/src/completed_homeworks/lazy_loading_completed_homeworks_bloc/lazy_loading_completed_homeworks_bloc.dart';
-import 'package:hausaufgabenheft_logik/src/completed_homeworks/lazy_loading_completed_homeworks_bloc/states.dart'
+import 'package:hausaufgabenheft_logik/src/completed_homeworks/lazy_loading_completed_homeworks_bloc/lazy_loading_completed_homeworks_bloc_impl.dart'
     as lazy_loading;
 import 'package:hausaufgabenheft_logik/src/completed_homeworks/views/completed_homework_list_view_factory.dart';
+import 'package:rxdart/rxdart.dart';
 
-import 'completed_homeworks_view_bloc.dart';
 import 'events.dart';
 import 'states.dart';
+
+export 'events.dart';
+export 'states.dart';
+
+abstract class CompletedHomeworksViewBloc
+    implements
+        Stream<CompletedHomeworksViewBlocState>,
+        Sink<CompletedHomeworksViewBlocEvent> {}
 
 class CompletedHomeworksViewBlocImpl extends Bloc<
         CompletedHomeworksViewBlocEvent, CompletedHomeworksViewBlocState>
     implements CompletedHomeworksViewBloc, BlocBase {
-  final LazyLoadingCompletedHomeworksBloc _lazyLoadingCompletedHomeworksBloc;
+  final lazy_loading.LazyLoadingCompletedHomeworksBloc
+      _lazyLoadingCompletedHomeworksBloc;
   final CompletedHomeworkListViewFactory _completedHomeworkListViewFactory;
   final int nrOfInitialCompletedHomeworksToLoad;
   StreamSubscription _streamSubscription;
