--- lib/hausaufgabenheft_logik/lib/src/completed_homeworks/lazy_loading_completed_homeworks_bloc/lazy_loading_completed_homeworks_bloc.dart
@@ -1,15 +1,0 @@
-// Copyright (c) 2022 Sharezone UG (haftungsbeschränkt)
-// Licensed under the EUPL-1.2-or-later.
-//
-// You may obtain a copy of the Licence at:
-// https://joinup.ec.europa.eu/software/page/eupl
-//
-// SPDX-License-Identifier: EUPL-1.2
-
-import 'events.dart';
-import 'states.dart';
-
-abstract class LazyLoadingCompletedHomeworksBloc
-    implements
-        Stream<LazyLoadingCompletedHomeworksBlocState>,
-        Sink<LazyLoadingCompletedHomeworksEvent> {}
